Extract decode.rs from encode.rs
